@@ -1,548 +1,255 @@
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:horizontal_calendar_widget/horizontal_calendar.dart';
+import 'package:horizontal_calendar_widget/date_helper.dart';
+import 'package:horizontal_calendar_widget/date_widget.dart';
 
 void main() {
-  testWidgets(
-    'Assert should fail if firstDate is NULL',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: null,
-                    lastDate: DateTime(2019, 11, 20),
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
+  final defaultLabelOrder = [
+    LabelType.month,
+    LabelType.date,
+    LabelType.weekday,
+  ];
 
   testWidgets(
-    'Assert should fail if lastDate is NULL',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: DateTime(2019, 11, 20),
-                    lastDate: null,
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
-
-  testWidgets(
-    'Assert should fail if Label order is NULL',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: DateTime(2019, 11, 20),
-                    lastDate: null,
-                    labelOrder: null,
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
-
-  testWidgets(
-    'Assert should fail if Label order is empty',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: DateTime(2019, 11, 20),
-                    lastDate: null,
-                    labelOrder: [],
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
-
-  testWidgets(
-    'Assert should fail if minSelectedDateCount is -gt maxSelectedDateCount',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: DateTime(2020, 03, 04),
-                    lastDate: DateTime(2020, 03, 08),
-                    minSelectedDateCount: 2,
-                    maxSelectedDateCount: 1,
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
-
-  testWidgets(
-    'Assert should fail if initialSelectedDate lenght is -lt minSelectedDateCount',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: DateTime(2020, 03, 04),
-                    lastDate: DateTime(2020, 03, 08),
-                    minSelectedDateCount: 2,
-                    maxSelectedDateCount: 3,
-                    initialSelectedDates: [
-                      DateTime(2020, 03, 04),
-                    ],
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
-
-  testWidgets(
-    'Assert should fail if initialSelectedDate lenght is -gt maxSelectedDateCount',
-    (WidgetTester tester) async {
-      expectLater(
-          () => tester.pumpWidget(
-                Directionality(
-                  child: HorizontalCalendar(
-                    firstDate: DateTime(2020, 03, 04),
-                    lastDate: DateTime(2020, 03, 08),
-                    maxSelectedDateCount: 2,
-                    initialSelectedDates: [
-                      DateTime(2020, 03, 04),
-                      DateTime(2020, 03, 05),
-                      DateTime(2020, 03, 06),
-                    ],
-                  ),
-                  textDirection: TextDirection.ltr,
-                ),
-              ),
-          throwsAssertionError);
-    },
-  );
-
-  testWidgets('Should render N widgets as provided start & end date',
+      'Render widgets with date / month/ weekday by default formats if formats are provided not explicitly',
       (WidgetTester tester) async {
     await tester.pumpWidget(
       Directionality(
-        child: HorizontalCalendar(
-          firstDate: DateTime(2019, 11, 17),
-          lastDate: DateTime(2019, 11, 20),
-          monthFormat: 'MMM',
-          weekDayFormat: 'EEE',
+        child: DateWidget(
+          date: DateTime(2019, 11, 17),
+          padding: const EdgeInsets.all(8),
+          labelOrder: defaultLabelOrder,
         ),
         textDirection: TextDirection.ltr,
       ),
     );
 
-    final month11 = find.text('Nov');
-    expect(month11, findsNWidgets(4));
-
     final date17 = find.text('17');
     expect(date17, findsOneWidget);
+    final month11 = find.text('Nov');
+    expect(month11, findsOneWidget);
     final week17 = find.text('Sun');
     expect(week17, findsOneWidget);
-
-    final date18 = find.text('18');
-    expect(date18, findsOneWidget);
-    final week18 = find.text('Mon');
-    expect(week18, findsOneWidget);
-
-    final date19 = find.text('19');
-    expect(date19, findsOneWidget);
-    final week19 = find.text('Tue');
-    expect(week19, findsOneWidget);
   });
 
   testWidgets(
-    'Default / Selected / Disabled widgets should render as per arguments',
+    'Render widgets with dates / month / week day by default formats if formats are provided explicitly',
     (WidgetTester tester) async {
-      final dateDecoration = BoxDecoration(
-        color: Colors.blue,
-      );
-      final selectedDateDecoration = BoxDecoration(
-        color: Colors.green,
-      );
-
-      final disabledDateDecoration = BoxDecoration(
-        color: Colors.grey,
-      );
-
       await tester.pumpWidget(
         Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            defaultDecoration: dateDecoration,
-            selectedDecoration: selectedDateDecoration,
-            initialSelectedDates: [DateTime(2019, 11, 18)],
-            disabledDecoration: disabledDateDecoration,
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
+          child: DateWidget(
+            date: DateTime(2019, 11, 17),
+            padding: const EdgeInsets.all(8),
+            dateFormat: 'dd/MMM',
+            monthFormat: 'MM',
+            weekDayFormat: 'EEEE',
+            labelOrder: defaultLabelOrder,
           ),
           textDirection: TextDirection.ltr,
         ),
       );
 
-      WidgetPredicate datePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == dateDecoration;
-      WidgetPredicate selectedDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == selectedDateDecoration;
-      WidgetPredicate disabledDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == disabledDateDecoration;
-
-      expect(find.byWidgetPredicate(datePredicate), findsOneWidget);
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsOneWidget);
-      expect(find.byWidgetPredicate(disabledDatePredicate), findsOneWidget);
+      final month11 = find.text('11');
+      expect(month11, findsOneWidget);
+      final date17 = find.text('17/Nov');
+      expect(date17, findsOneWidget);
+      final week17 = find.text('Sunday');
+      expect(week17, findsOneWidget);
     },
   );
 
   testWidgets(
-    'State should change after tapping any of date',
+    'Default decoration should be applied to Date / month / weekday if not selected',
     (WidgetTester tester) async {
-      final dateDecoration = BoxDecoration(
+      const dateDecoration = BoxDecoration(
         color: Colors.blue,
-      );
-      final selectedDateDecoration = BoxDecoration(
-        color: Colors.green,
-      );
-
-      final disabledDateDecoration = BoxDecoration(
-        color: Colors.grey,
       );
 
       await tester.pumpWidget(
         Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
+          child: DateWidget(
+            date: DateTime(2019, 11, 17),
+            padding: const EdgeInsets.all(8),
+            dateFormat: null,
+            monthFormat: null,
+            weekDayFormat: null,
             defaultDecoration: dateDecoration,
-            selectedDecoration: selectedDateDecoration,
-            initialSelectedDates: [DateTime(2019, 11, 18)],
-            disabledDecoration: disabledDateDecoration,
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-            maxSelectedDateCount: 2,
+            labelOrder: defaultLabelOrder,
           ),
           textDirection: TextDirection.ltr,
         ),
       );
 
-      WidgetPredicate datePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == dateDecoration;
-      WidgetPredicate selectedDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == selectedDateDecoration;
-      WidgetPredicate disabledDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == disabledDateDecoration;
-
-      expect(find.byWidgetPredicate(datePredicate), findsOneWidget);
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsOneWidget);
-      expect(find.byWidgetPredicate(disabledDatePredicate), findsOneWidget);
-
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String())));
-
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsNWidgets(2));
-
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String())));
-
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsOneWidget);
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Container && widget.decoration == dateDecoration),
+          findsOneWidget);
     },
   );
 
   testWidgets(
-    '''Case: maxSelectedDateCount = 1. 
-    only one date should be selected at a time. 
-    On select new date, previous date should be unselected''',
+    'Selected decoration should be applied to Date / month / weekday is selected',
     (WidgetTester tester) async {
-      final dateDecoration = BoxDecoration(
+      const dateDecoration = BoxDecoration(
         color: Colors.blue,
-      );
-      final selectedDateDecoration = BoxDecoration(
-        color: Colors.green,
-      );
-
-      final disabledDateDecoration = BoxDecoration(
-        color: Colors.grey,
       );
 
       await tester.pumpWidget(
         Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            defaultDecoration: dateDecoration,
-            selectedDecoration: selectedDateDecoration,
-            disabledDecoration: disabledDateDecoration,
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-            maxSelectedDateCount: 1,
+          child: DateWidget(
+            date: DateTime(2019, 11, 17),
+            padding: const EdgeInsets.all(8),
+            dateFormat: null,
+            monthFormat: null,
+            weekDayFormat: null,
+            selectedDecoration: dateDecoration,
+            isSelected: true,
+            labelOrder: defaultLabelOrder,
           ),
           textDirection: TextDirection.ltr,
         ),
       );
 
-      WidgetPredicate selectedDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == selectedDateDecoration;
-
-      //Tap date 17 and expect only single date should be selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsOneWidget);
-
-      //Again tap 17 and expect it should be un selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsNothing);
-
-      //Tap 17 and then 18 and expect only one date should be selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String())));
-      await tester.pumpAndSettle();
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 18).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsOneWidget);
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Container && widget.decoration == dateDecoration),
+          findsOneWidget);
     },
   );
 
   testWidgets(
-    '''Case: maxSelectedDateCount = 2. 
-    only two dates should be selected at a time. 
-    If max count reached, new dates should not be selected''',
+    'Disabled decoration should be applied to Date / month / weekday if is disabled',
     (WidgetTester tester) async {
-      final dateDecoration = BoxDecoration(
+      const dateDecoration = BoxDecoration(
         color: Colors.blue,
       );
-      final selectedDateDecoration = BoxDecoration(
-        color: Colors.green,
-      );
 
-      final disabledDateDecoration = BoxDecoration(
-        color: Colors.grey,
-      );
+      const style = TextStyle(color: Colors.white);
 
       await tester.pumpWidget(
         Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            defaultDecoration: dateDecoration,
-            selectedDecoration: selectedDateDecoration,
-            disabledDecoration: disabledDateDecoration,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-            maxSelectedDateCount: 2,
+          child: DateWidget(
+            date: DateTime(2019, 11, 17),
+            padding: const EdgeInsets.all(8),
+            dateFormat: null,
+            monthFormat: null,
+            weekDayFormat: null,
+            disabledDecoration: dateDecoration,
+            isDisabled: true,
+            dateTextStyle: style,
+            labelOrder: defaultLabelOrder,
           ),
           textDirection: TextDirection.ltr,
         ),
       );
 
-      WidgetPredicate selectedDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == selectedDateDecoration;
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Container && widget.decoration == dateDecoration),
+          findsOneWidget);
 
-      //Tap 17 and then 18 and expect two dates should be selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String())));
-      await tester.pumpAndSettle();
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 18).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsNWidgets(2));
-
-      //Tap 19 and expect two dates should be selected: 19 should not be selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2019, 11, 19).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsNWidgets(2));
+      expect(
+          find.byWidgetPredicate(
+              (Widget widget) => widget is Text && widget.style == style),
+          findsOneWidget);
     },
   );
 
   testWidgets(
-    '''Case: minSelectedDateCount = 1 and maxSelectedDateCount = 1. 
-    one and only one date should be always selected.''',
+    'Default Date with all properties',
     (WidgetTester tester) async {
-      final dateDecoration = BoxDecoration(
+      const defaultDecoration = BoxDecoration(
         color: Colors.blue,
       );
-      final selectedDateDecoration = BoxDecoration(
-        color: Colors.green,
-      );
 
-      final disabledDateDecoration = BoxDecoration(
-        color: Colors.grey,
-      );
+      const padding = EdgeInsets.symmetric(horizontal: 8);
+
+      const monthStyle = TextStyle(color: Colors.white);
+      const monthFormat = 'MM';
+
+      const dateStyle = TextStyle(color: Colors.black);
+      const dateFormat = 'dd/MM';
+
+      const weekDayStyle = TextStyle(color: Colors.green);
+      const weekDayFormat = 'EEE';
 
       await tester.pumpWidget(
         Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2020, 03, 04),
-            lastDate: DateTime(2020, 03, 08),
-            defaultDecoration: dateDecoration,
-            selectedDecoration: selectedDateDecoration,
-            disabledDecoration: disabledDateDecoration,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-            minSelectedDateCount: 1,
-            maxSelectedDateCount: 1,
-            initialSelectedDates: [
-              DateTime(2020, 03, 05),
+          child: DateWidget(
+            date: DateTime(2019, 11, 17),
+            padding: padding,
+            dateFormat: dateFormat,
+            dateTextStyle: dateStyle,
+            monthFormat: monthFormat,
+            monthTextStyle: monthStyle,
+            weekDayFormat: weekDayFormat,
+            weekDayTextStyle: weekDayStyle,
+            defaultDecoration: defaultDecoration,
+            labelOrder: defaultLabelOrder,
+          ),
+          textDirection: TextDirection.ltr,
+        ),
+      );
+
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Container && widget.decoration == defaultDecoration),
+          findsOneWidget);
+
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Text &&
+              widget.style == dateStyle &&
+              widget.data == '17/11'),
+          findsOneWidget);
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Text &&
+              widget.style == monthStyle &&
+              widget.data == '11'),
+          findsOneWidget);
+
+      expect(
+          find.byWidgetPredicate((Widget widget) =>
+              widget is Text &&
+              widget.style == weekDayStyle &&
+              widget.data == 'Sun'),
+          findsOneWidget);
+    },
+  );
+
+  testWidgets(
+    'Lables should render as per provided order',
+    (WidgetTester tester) async {
+      await tester.pumpWidget(
+        Directionality(
+          child: DateWidget(
+            date: DateTime(2019, 11, 17),
+            padding: const EdgeInsets.all(8),
+            labelOrder: const [
+              LabelType.date,
+              LabelType.month,
             ],
+            dateTextStyle: null,
           ),
           textDirection: TextDirection.ltr,
         ),
       );
 
-      WidgetPredicate selectedDatePredicate = (Widget widget) =>
-          widget is Container && widget.decoration == selectedDateDecoration;
-
-      //Tap 04 and expect one date still being selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2020, 03, 04).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsNWidgets(1));
-
-      //Tap 05 and expect still only one date being selected
-      await tester
-          .tap(find.byKey(Key(DateTime(2020, 03, 05).toIso8601String())));
-      await tester.pumpAndSettle();
-      expect(find.byWidgetPredicate(selectedDatePredicate), findsNWidgets(1));
-    },
-  );
-
-  testWidgets(
-    'onDateSelected callback should be invoked',
-    (WidgetTester tester) async {
-      await tester.pumpWidget(
-        Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            initialSelectedDates: [DateTime(2019, 11, 18)],
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-          ),
-          textDirection: TextDirection.ltr,
-        ),
-      );
-
-      expectLater(
-        () => tester
-            .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String()))),
-        prints('S:${DateTime(2019, 11, 17)}\n'),
-      );
-    },
-  );
-
-  testWidgets(
-    'onDateUnSelected callback should be invoked',
-    (WidgetTester tester) async {
-      await tester.pumpWidget(
-        Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            initialSelectedDates: [DateTime(2019, 11, 18)],
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-          ),
-          textDirection: TextDirection.ltr,
-        ),
-      );
-
-      expectLater(
-        () => tester
-            .tap(find.byKey(Key(DateTime(2019, 11, 18).toIso8601String()))),
-        prints('U:${DateTime(2019, 11, 18)}\n'),
-      );
-    },
-  );
-
-  testWidgets(
-    'onDateLongTap callback should be invoked',
-    (WidgetTester tester) async {
-      await tester.pumpWidget(
-        Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            initialSelectedDates: [DateTime(2019, 11, 18)],
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
-            onDateSelected: (date) => print("S:$date"),
-            onDateUnSelected: (date) => print("U:$date"),
-            onDateLongTap: (date) => print("L:$date"),
-          ),
-          textDirection: TextDirection.ltr,
-        ),
-      );
-
-      expectLater(
-        () => tester.longPress(
-            find.byKey(Key(DateTime(2019, 11, 18).toIso8601String()))),
-        prints('L:${DateTime(2019, 11, 18)}\n'),
-      );
-    },
-  );
-
-  testWidgets(
-    'onMaxDateSelectionReached callback should be invoked',
-    (WidgetTester tester) async {
-      await tester.pumpWidget(
-        Directionality(
-          child: HorizontalCalendar(
-            firstDate: DateTime(2019, 11, 17),
-            lastDate: DateTime(2019, 11, 19),
-            isDateDisabled: (date) =>
-                date.compareTo(DateTime(2019, 11, 19)) == 0,
-            onMaxDateSelectionReached: () => print("MAX_SELCTION"),
-            maxSelectedDateCount: 0,
-          ),
-          textDirection: TextDirection.ltr,
-        ),
-      );
-
-      expectLater(
-        () => tester
-            .tap(find.byKey(Key(DateTime(2019, 11, 17).toIso8601String()))),
-        prints('MAX_SELCTION\n'),
-      );
+      expect(find.byWidgetPredicate((Widget widget) {
+        if (widget is Column) {
+          final children = widget.children;
+          if (children.length < 2) {
+            return false;
+          }
+          Text firstText = children[0] as Text;
+          Text secondText = children[1] as Text;
+          if (firstText.data == "17" && secondText.data == "Nov") {
+            return true;
+          }
+        }
+        return false;
+      }), findsOneWidget);
     },
   );
 }
